IPv6 support for all modules (#64)
* IPv6 and IP protocol preferrence support for all modules

Introduce protocol field for every module which can be used to force
probe to IPv4 (tcp4/udp4/icmp4) or IPv6 (tcp6/udp6/icmp6). In case
of tcp/udp/icmp both can be used in order of preferrence. IPv6 is
preferred by default, but it can be tuned with 'preferred_ip_protocol'
parameter.

Default for all modules is an automatic selection with IPv6 preferred.

ICMP code is mostly from Michael Stapelberg.

* Document protocol and preferred_ip_protocol parameters

* Protocol tests for DNS and TCP modules