Change the front page links to probe logs into relative links
The front page contains a table of links to logs from the most recent
probes. These links were accidentally changed to absolute links in
252a878, in a form that doesn't work when using --web.external-url.
Rather than prefixing these with the external URL (if set), we return
them to being relative links, as they are meant to be (per issue #519).

Fixes #519

Signed-off-by: Chris Siebenmann <cks.git01@cs.toronto.edu>