prober/tls: fix probe_ssl_last_chain_expiry_timestamp_seconds
This metric should report the earliest expiry of the chain that expires
the latest out of all the verified chains. Presently, it reports the
earliest expiry of the chain that expires first.

The current test for this metric was using an expired root certificate which
is omitted from the verified chain, so the test was passing despite this
bug. I've changed it to use a root that is still valid but expires before a
root held by the client.

Signed-off-by: Rob Best <robertbest89@gmail.com>