Use DialContext to deal with pre-resolving for HTTP
This replaces the method of resolving the IP and then munging the URL,
with essentially the other way around. It puts the desired hostname in
the URL, then uses a DialContext function to override the particular
host that is specified in target.

To try to explain that better, this behaves more like curl does when:
`curl --resolve example.com:443:192.0.2.1 http://example.com` is used.
That is, any request to the host specified (such as via redirects) will
continue to use the resolved IP.

This solves a very hard to debug problem -- if a target sends a
redirect, which has its own hostname in it, then blackbox_exporter will
look up that hostname via DNS and connect to it. This means the second
redirect could hit a different server.

An additional upside of this is the TLS server name logic is not needed
-- we are giving the Go HTTP client the real hostname, so it can check
SNIs for us (e.g. this solves some issues around port numbers in SNIs,
which blackbox_exporter didn't strip from the Host header).

Further, this deliberately makes no assumptions about other ports, so
for example a HTTP -> HTTPS redirect will result in a DNS resolution. It
also adds a log for the actual IP and port being connected to, making
debugging easier, and removing the confusing log message where the URL
is replaced with the IP.

Signed-off-by: David Leadbeater <dgl@dgl.cx>