http: Prevent logging confusing error message
When FailIfNotSSL is set, the error message:

> Final request was not over SSL

is logged even if a previous error has prevented the request to
complete, for example when it fails due to a timeout or connection
error.

This updates the logic around the log message so it's printed when an
otherwise successful request is not performed over SSL and that causes
the probe to fail.

Signed-off-by: Adrian Serrano <adrisr83@gmail.com>