fix: Correct tool existence checks
`command -v <cmd>` returns a path to a binary to stdout if it's found in PATH.
The condition deciding if `<cmd>` is installed redirects stdout to `/dev/null`
and so this check could never work.

This change redirects stderr to `/dev/null` instead of stdout to compare the
output of `command -v <cmd>` properly with an empty string.

I unified the redirection to a file in whole file. The empty space behind
redirection is not needed.

Signed-off-by: Pavel Pulec <pulecp@gmail.com>