fix: stop scrape logger spam
Immediate fix for #1377. This fix will need to be refactored a bit when
we upgrade prometheus/common to v0.63.0, since prometheus/common#754
contains a breaking change to promslog's level code.

Explanation:
When I ported the blackbox exporter to use slog in #1311, I removed the
`none` option from the log prober flag, as it was non-standard to other
prometheus projects, not a valid value in the original promlog package,
and slog has no inherent corresponding concept. The log prober flag
value is used to create the scrape logger (which is specifically used
for scrape-related logging, not overall application logging). I also
included some code that ensured a level config would be created for use
with the scrape logger. In combination, this causes the blackbox
exporter to always output scrape logs, which can be noisy. If we instead
default the flag to an empty string and do some intelligent handling of
an nil log prober value within the scrapeLogger's implementation of the
slog.Handler interface, then we can restore the previous behavior.

TL;DR:
Leave `--log.prober` flag unset and it won't do scrape-level logging

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>